+++ docs/src/examples/components/Button/Types/ButtonExampleContentAndIcon.shorthand.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+import { Button } from '@stardust-ui/react'
+
+const ButtonExampleContentAndIcon = () => (
+  <div>
+    <Button type="primary" icon="book" content="Click me left" iconPosition="left" />
+    <Button type="secondary" icon="coffee" content="Click me right" iconPosition="right" />
+  </div>
+)
+
+export default ButtonExampleContentAndIcon